+++ Sources/ProjectSpec/SpecParsingError.swift
@@ -6,6 +6,7 @@ public enum SpecParsingError: Error, CustomStringConvertible {
     case invalidDependency([String: Any])
     case unknownPackageRequirement([String: Any])
     case invalidSourceBuildPhase(String)
+    case invalidTargetReference(String)
     case invalidVersion(String)
 
     public var description: String {
@@ -18,6 +19,8 @@ public enum SpecParsingError: Error, CustomStringConvertible {
             return "Unknown Target dependency: \(dependency)"
         case let .invalidSourceBuildPhase(error):
             return "Invalid Source Build Phase: \(error)"
+        case let .invalidTargetReference(targetReference):
+            return "Invalid Target Reference Syntax: \(targetReference)"
         case let .invalidVersion(version):
             return "Invalid version: \(version)"
         case let .unknownPackageRequirement(package):