+++ Tests/ProjectSpecTests/SpecLoadingTests.swift
@@ -2,10 +2,10 @@ import Foundation
 import PathKit
 import ProjectSpec
 import Spectre
-import XcodeGenKit
 import XcodeProj
 import XCTest
 import Yams
+import TestSupport
 
 class SpecLoadingTests: XCTestCase {
 
@@ -467,7 +467,7 @@ class SpecLoadingTests: XCTestCase {
                             "platform": "iOS",
                             "sources": [
                                 "templateSource",
-                                ["path": "Sources/${target_name}"]
+                                ["path": "Sources/${target_name}"],
                             ],
                         ],
                         "temp2": [
@@ -750,6 +750,7 @@ class SpecLoadingTests: XCTestCase {
                             "Target4": ["testing": true],
                             "Target5": ["testing": false],
                             "Target6": ["test", "analyze"],
+                            "ExternalProject/Target7": ["run"],
                         ],
                         "preActions": [
                             [
@@ -764,7 +765,7 @@ class SpecLoadingTests: XCTestCase {
                         "targets": [
                             "Target1",
                             [
-                                "name": "Target2",
+                                "name": "ExternalProject/Target2",
                                 "parallelizable": true,
                                 "randomExecutionOrder": true,
                                 "skippedTests": ["Test/testExample()"],
@@ -782,6 +783,7 @@ class SpecLoadingTests: XCTestCase {
                     Scheme.BuildTarget(target: "Target4", buildTypes: [.testing]),
                     Scheme.BuildTarget(target: "Target5", buildTypes: []),
                     Scheme.BuildTarget(target: "Target6", buildTypes: [.testing, .analyzing]),
+                    Scheme.BuildTarget(target: "ExternalProject/Target7", buildTypes: [.running]),
                 ]
                 try expect(scheme.name) == "Scheme"
                 try expect(scheme.build.targets) == expectedTargets
@@ -799,7 +801,7 @@ class SpecLoadingTests: XCTestCase {
                     targets: [
                         "Target1",
                         Scheme.Test.TestTarget(
-                            name: "Target2",
+                            targetReference: "ExternalProject/Target2",
                             randomExecutionOrder: true,
                             parallelizable: true,
                             skippedTests: ["Test/testExample()"]
@@ -872,7 +874,7 @@ class SpecLoadingTests: XCTestCase {
                             "platform": "iOS",
                             "sources": [
                                 "templateSource",
-                                ["path": "Sources/${target_name}"]
+                                ["path": "Sources/${target_name}"],
                             ],
                         ],
                         "temp2": [
@@ -960,7 +962,7 @@ class SpecLoadingTests: XCTestCase {
                     targets: [
                         "TargetFirstTarget",
                         Scheme.Test.TestTarget(
-                            name: "Target2",
+                            targetReference: "Target2",
                             randomExecutionOrder: true,
                             parallelizable: true,
                             skippedTests: ["Test/testExample()"]
@@ -1087,14 +1089,14 @@ class SpecLoadingTests: XCTestCase {
                     "package5": SwiftPackage(url: "package.git", versionRequirement: .revision("x")),
                     "package6": SwiftPackage(url: "package.git", versionRequirement: .range(from: "1.2.0", to: "1.2.5")),
                     "package7": SwiftPackage(url: "package.git", versionRequirement: .exact("1.2.2")),
-                    ],
-                    localPackages: ["../../Package"],
-                    options: .init(localPackagesGroup: "MyPackages"))
+                ],
+                                      localPackages: ["../../Package"],
+                                      options: .init(localPackagesGroup: "MyPackages"))
 
                 let dictionary: [String: Any] = [
                     "name": "spm",
                     "options": [
-                        "localPackagesGroup": "MyPackages"
+                        "localPackagesGroup": "MyPackages",
                     ],
                     "packages": [
                         "package1": ["url": "package.git", "exactVersion": "1.2.2"],
@@ -1105,7 +1107,7 @@ class SpecLoadingTests: XCTestCase {
                         "package6": ["url": "package.git", "minVersion": "1.2.0", "maxVersion": "1.2.5"],
                         "package7": ["url": "package.git", "version": "1.2.2"],
                     ],
-                    "localPackages": ["../../Package"]
+                    "localPackages": ["../../Package"],
                 ]
                 let parsedSpec = try getProjectSpec(dictionary)
                 try expect(parsedSpec) == project